Improvement: Make rest-gateway resources configurable
Depending on the environment this is running we dont really need to
request half of a cpu. Make this configurable to be more flexible.

Furthermore I moved the default resources to ask for request instead of
limit. Limiting the pod in cpu in a bit of an anti pattern on
kubernetes.