Bugfix: Reuse existing secrets
With the current implementation we would have to choice between creating
the secret manually or have it auto created but then it will be
recreated at each deploy.

With this change. If the secret is already present the values of the
secret will be reused, otherwise the secret will be created with either
the values configured or some random generated values.

This way the secret will only be created if it does not exist.